scheduler/requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
coloredlogs==15.0.1
DateTime==4.4
dnspython==2.2.1
humanfriendly==10.0
idna==3.3
pandas
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
urllib3==1.26.9
zope.interface==5.4.0
openpyxl==3.0.10
mysql-connector-python==8.0.28
dataclasses==0.6
dataclasses-serialization==1.3.1
boto3
mongoengine==0.27.0
attrs==23.1.0
zipp==3.15.0
zope.interface==5.4.0
-e ./share
beautifulsoup4
pycountry
redis~=4.4.0
xmltodict
djangorestframework==3.14.0
pydantic
paramiko
